Update for 2025
お待たせしました。
Fixed some bugs with the wizard and changed some functionality so things are easier to understand.
Added a rough Japanese translation for most sections of the program if your language is set to Japanese.
Updated MML version and included the relevant files within the repo for easier self-building.
If you have any issues please report them here.